requirements.txt
django==5.1.1
environs==11.0.0
psycopg2-binary==2.9.5
pillow==11.0.0
phonenumbers==8.13.48
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
django-debug-toolbar==4.4.6